:recycle: (client/settings/auth) Refactor Authentication component for settings
