+++ src/components/settings/auth/SecurityNote.tsx
@@ -1,0 +1,9 @@
+export default function SecurityNote() {
+  return (
+    <div className="mt-4 p-3 bg-gray-50 dark:bg-gray-900 rounded-md border border-gray-200 dark:border-gray-700">
+      <p className="text-xs text-gray-500 dark:text-gray-400">
+        <strong>Security Note:</strong> For enhanced security, ImageWarden will automatically log you out after 30 minutes of inactivity.
+      </p>
+    </div>
+  );
+}
